Adams Code emailed on saturday august 5th
As I said on Wed, I've tried to put together some sketch code that
might be helpful for the group as you begin to build the Processing
visualization. What I have working is a Processing sketch that imports
three elements of a Jokbo CSV (year, lat, long), which I cut down and
exported myself. This data gives us the event year and the lat & long
the bride came from. The sketch is OO and uses a class called
Adventure which takes as input the year, original x & y, destination x
& y. The class has a method for moving an ellipse slightly towards the
destination each time it is called (60 times a second under default
Processing FPS). I've limited it by year in a very rough way (static
int for upper and lower bounds) but this should not be difficult to
automate. What it would require, in my thinking, is some way to know
all of the bride objects have arrived (probably via some property of
the Adventure class?). It would also be possible to change the
Adventure class so that every bride arrived by some pre-determined
time point so that, e.g. every bride from 1900 - 1925 arrives over the
course of 10 minutes. This would require dynamically changing the
scaling factor in the Adventure class (currently set to a static .1).
So close-by brides would arrive early and the furthers bride would
arrive last. This would require going through the file once before
anyone started moving, finding the furtherest bride(s) and mapping
(http://processing.org/reference/map_.html) the scaling factor for
each bride object based on the time. I think that makes sense but
honestly, I'm done for the day :)

Hopefully this framework gives you a good start towards moving
multiple data points towards a common location (my understanding is
that you will have brides moving towards a single clan). I was not
able to address the Korea map SVG (I am fairly new to vector math, so
that took most of the day). It's something other people have done so I
am sure with Miguel's help we can get that working. The speed of the
entire thing is definitely affected by the number of objects, I think
there may me several redundant objects which cannot even be seen.
These could be culled from the data beforehand or not displayed or,
more complex, tallied before display with a corresponding increase in
ellipse (etc) size.